// libs/shared/step-flow/src/lib/interfaces/step-config.interface.ts
import { StepLayoutOptionsInterface } from '../step-layout-options.interface';
import { StepAnalyticsOptions } from './step-analytics-options.interface';

export interface StepConfig {
  key: string;
  imagePath?: string;
  header?: string;
  component: unknown;
  isFinalState?: boolean;
  headerDescription?: boolean;
  headerSubDescription?: boolean;

  layoutOptions?: StepLayoutOptionsInterface;
  analyticsOptions?: StepAnalyticsOptions;
}